Address PR review feedback for dependabot automerge
- Guard against None in _get_repo_from_event with explicit isinstance
- Add full NumPy-style docstring to AutomergeOptions dataclass
- Extract GraphQL client logic into separate graphql_client module
- Use type keyword for JsonValue type alias (PEP 695)
- Simplify _emit_decision by hoisting duplicated reason assignment
- Catch httpx.TransportError instead of broader HTTPError
