+++ workflow_scripts/dependabot_automerge.py
@@ -61,18 +61,14 @@ import dataclasses
 import json
 import os
 import sys
-import time
 import typing as typ
 from pathlib import Path
 
-import httpx
 from cyclopts import App, Parameter
 
-# Type alias for JSON-compatible values (parsed from json.loads)
-JsonValue = str | int | float | bool | None | list["JsonValue"] | dict[str, "JsonValue"]
+from workflow_scripts.graphql_client import JsonValue, request_graphql
 
 DEPENDABOT_LOGIN = "dependabot[bot]"
-GRAPHQL_ENDPOINT = "https://api.github.com/graphql"
 
 MERGE_METHODS = {
     "merge": "MERGE",
@@ -266,9 +262,12 @@ def _get_repo_from_event(event: dict[str, JsonValue] | None) -> str:
     """Return the repository full name from an event payload when available."""
     if not event:
         return ""
-    repo = event.get("repository", {}).get("full_name")
-    if isinstance(repo, str):
-        return repo.strip()
+    repository = event.get("repository")
+    if not isinstance(repository, dict):
+        return ""
+    full_name = repository.get("full_name")
+    if isinstance(full_name, str):
+        return full_name.strip()
     return ""
 
 
@@ -373,12 +372,11 @@ def _emit_decision(
     config: AutomergeConfig,
 ) -> None:
     """Emit structured decision output for the automerge workflow."""
+    reason = decision.reason
     if decision.status == "ready" and config.dry_run:
         status = "dry-run"
-        reason = decision.reason
     else:
         status = decision.status
-        reason = decision.reason
     _emit("automerge_status", status)
     _emit("automerge_reason", reason)
     _emit("automerge_merge_method", config.merge_method)
@@ -388,98 +386,6 @@ def _emit_decision(
     _emit("automerge_author", pr.author)
     _emit("automerge_draft", str(pr.is_draft).lower())
     _emit("automerge_labels", pr.labels)
-
-
-def _should_retry(attempt: int, max_retries: int) -> bool:
-    """Return True if more retry attempts remain."""
-    return attempt < max_retries
-
-
-def _backoff_sleep(attempt: int, base_seconds: float = 1.0) -> None:
-    """Sleep with exponential backoff based on attempt number."""
-    time.sleep(base_seconds * (2**attempt))
-
-
-def _attempt_retry_or_fail(attempt: int, max_retries: int, error_message: str) -> None:
-    """Sleep for backoff if retries remain, otherwise fail with the given message."""
-    if not _should_retry(attempt, max_retries):
-        _fail(error_message)
-    _backoff_sleep(attempt)
-
-
-def _parse_graphql_response(response: httpx.Response) -> dict[str, JsonValue]:
-    """Parse and validate a GraphQL response, returning the data payload."""
-    try:
-        payload = response.json()
-    except json.JSONDecodeError as exc:
-        _fail(f"GitHub API response was not valid JSON: {exc}")
-
-    errors = payload.get("errors")
-    if errors:
-        _fail(f"GitHub API GraphQL errors: {errors}")
-
-    data = payload.get("data")
-    if data is None:
-        _fail("GitHub API returned no data.")
-    return data
-
-
-def _execute_graphql_attempt(
-    token: str, query: str, variables: dict[str, JsonValue]
-) -> httpx.Response | None:
-    """Execute a single GraphQL request attempt, returning None on connection error."""
-    headers = {
-        "Authorization": f"Bearer {token}",
-        "Accept": "application/vnd.github+json",
-    }
-    try:
-        with httpx.Client(timeout=30) as client:
-            return client.post(
-                GRAPHQL_ENDPOINT,
-                json={"query": query, "variables": variables},
-                headers=headers,
-            )
-    except httpx.HTTPError:
-        return None
-
-
-def _handle_response_or_retry(
-    response: httpx.Response | None, attempt: int, max_retries: int
-) -> httpx.Response | None:
-    """Process response or decide to retry. Returns None to signal retry."""
-    if response is None:
-        _attempt_retry_or_fail(
-            attempt,
-            max_retries,
-            "GitHub API request failed after retries: connection error",
-        )
-        return None
-
-    if 400 <= response.status_code < 500:
-        _fail(f"GitHub API error {response.status_code}: {response.text}")
-
-    if response.status_code >= 500:
-        msg = f"GitHub API error {response.status_code} after retries"
-        _attempt_retry_or_fail(attempt, max_retries, f"{msg}: {response.text}")
-        return None
-
-    return response
-
-
-def _request_graphql(
-    token: str, query: str, variables: dict[str, JsonValue]
-) -> dict[str, JsonValue]:
-    """Execute a GraphQL request with retry logic and return the data payload."""
-    max_retries = 3
-
-    for attempt in range(max_retries + 1):
-        response = _execute_graphql_attempt(token, query, variables)
-        processed_response = _handle_response_or_retry(response, attempt, max_retries)
-        if processed_response is None:
-            continue
-        return _parse_graphql_response(processed_response)
-
-    _fail("GitHub API request failed unexpectedly.")
 
 
 def _extract_author_login(pull_request: dict[str, JsonValue]) -> str:
@@ -515,7 +421,7 @@ def _fetch_pull_request(
     token: str, owner: str, repo: str, number: int
 ) -> PullRequestContext:
     """Fetch PR metadata from the GitHub GraphQL API."""
-    data = _request_graphql(
+    data = request_graphql(
         token,
         PULL_REQUEST_QUERY,
         {"owner": owner, "name": repo, "number": number},
@@ -544,7 +450,7 @@ def _fetch_pull_request(
 
 def _enable_automerge(token: str, pull_request_id: str, merge_method: str) -> None:
     """Enable auto-merge on a pull request via the GitHub GraphQL API."""
-    _request_graphql(
+    request_graphql(
         token,
         ENABLE_AUTOMERGE_MUTATION,
         {"pullRequestId": pull_request_id, "mergeMethod": merge_method},
@@ -613,7 +519,22 @@ def _handle_live_execution(
 
 @dataclasses.dataclass(frozen=True, slots=True)
 class AutomergeOptions:
-    """CLI options for automerge execution."""
+    """CLI options for automerge execution.
+
+    Attributes
+    ----------
+    merge_method : str
+        Merge method to use: ``squash``, ``merge``, or ``rebase``.
+    required_label : str or None
+        Label that must be present on the PR, or None to skip label checks.
+    dry_run : bool
+        If True, logs the decision without calling the GitHub API.
+    pull_request_number : int or None
+        Explicit PR number override, or None to resolve from event payload.
+    repository : str or None
+        Repository override in ``owner/repo`` form, or None to resolve from
+        event payload or environment.
+    """
 
     merge_method: typ.Annotated[
         str,
